pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["bloodywd <bloodyforreal@gmail.com>"]
readme = "README.md"
packages = [
  { include = "task_manager" },
]
classifiers = [
  "Environment :: Web Environment",
  "Framework :: Django",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Topic :: Internet :: WWW/HTTP :: Site Management :: Link Checking",
]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
django-bootstrap5 = "^23.4"
django-filter = "^24.2"
rollbar = "^0.16.3"
django-extensions = "^3.2.3"
pytest-cov = "^5.0.0"
coverage = "^7.6.1"
flake8 = "^7.1.1"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
django-extensions = "^3.2.3"
ipython = "^8.22.2"
dj-database-url = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
